restructured the api to seperate concerns
