+++ api/routers/routes.py
@@ -1,6 +1,8 @@
 from typing import Any, Optional
 import asyncpg
 from fastapi import APIRouter, Request
+
+
 router = APIRouter()
 
 
@@ -15,3 +17,18 @@ async def current_time(request: Request) -> dict[str, Any]:
         return {"now": now.isoformat()}
     except AttributeError:
         return {"now": str(now)}
+
+@router.get("/health")
+async def health_check(request: Request) -> dict[str, Any]:
+    pool: Optional[asyncpg.Pool] = getattr(request.app.state, "db_pool", None)
+    if not pool:
+        err = getattr(request.app.state, "db_error", None)
+        if err:
+            return {"status": "degraded", "db": "error", "error": err}
+        return {"status": "ok", "db": "not_configured"}
+    try:
+        async with pool.acquire() as conn:
+            val = await conn.fetchval("SELECT 1")
+        return {"status": "ok", "db": "up", "result": int(val)}
+    except Exception as exc:
+        return {"status": "degraded", "db": "error", "error": str(exc)}
